add kills to variables declared in the same scope that holds the scop
In particular, if a variable is declared in the same scope,
does not have its address taken and is not used after the scop,
then the variable should not be considered live-out with respect
to the scop.  Add kills to these variables, provided they also
appear somewhere inside the scop, such that they do not need
to be considered live-out.

This is useful for some of the PolyBench benchmarks, which
have temporary variables declared outside of the marked region.
Without the kills, the user would have to assume that those
variable may be live-out, reducing the set of possible transformations.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>