update test case outputs
pet_scop_cmp purposely tries not to depend too much on the exact
output produced by isl so that we do not have to update the test
case outputs each time isl is updated.  However, to ease a visual
inspection of subsequent changes to the pet output, we do perform
such an update now.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>