pet_expr_extract_affine: cache result in pet_context
pet_expr_extract_affine may be called several times on the same
pet_expr within the same context.  This is even more likely to
happen after changing pet to also look for affine expressions
outside accesses.
Store the result in the pet_context for later use.
The results are invalidated whenever the pet_context is changed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>