push affine conditions into index expressions
If the code contains an expression of the form

    c ? A[f] : A[g]

with c an affine condition, then replace it by

    A[c ? f : g]

This reduces the number of accesses and, more importantly,
includes the information of when the access is performed
into the index expression.  This allows dependence analysis
to take into account this information, reducing the risk of
introducing spurious dependences.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>