remove "const" qualifier from the type of a scalar
A scalar that is qualified "const" may still be initialized inside the code.
Users of pet, e.g., PPCG, may need to break up the declaration
with initialization into a pure declaration and an assignment.
Such assignments are not allowed on variables of "const" type.
Since it is easier to remove the "const" qualifier in pet,
do it here.  This solves a problem in PPCG reported by
Bruno bodin <bbodin@inf.ed.ac.uk>.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>