perform substitutions on size expression of array declared in inlined function
During inlining, the function parameters may get renamed apart from
other variables in the calling function.  These renamings are applied
to the tree representing the body of the inlined function, but
the size expressions of the arrays declared in the function are
only derived afterwards and these substitutions are therefore not
taken into account in the size expressions.
Precompute these size expressions and replace them by the result
of applying the substitutions such that later calls to get_array_size
will return the modified size expressions.

Reported-by: Bruno bodin <bbodin@inf.ed.ac.uk>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>