tree2scop.c: tree_is_decl: do not treat empty block as group of declarations
A block was being treated as a group of declarations if all
its elements are declarations, including the case where
the block has no elements.  Do not treat such empty blocks
as a group of declarations.

This improves support for empty statements since a block of code
consisting of only empty statements results in an empty pet_tree_block,
which would then mistakenly get treated as a group of declarations.

Reported-by: Samuel Thibault <samuel.thibault@inria.fr>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>