pet_array: explicitly keep track of whether array is innermost
pet_scop_compute_*_to_inner uses the element_is_record attribute
of a pet_array to determine whether it is innermost.
This is correct in the current code base because any pet_array
that has a record element has its subfields represented as pet_array
objects as well.
In a future commit, only subfields that are relevant for the model
will be added to the list of pet_array objects.
The pet_array therefore needs to explicitly keep track of
whether the array is innermost in the model.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>