pet_scop_collect_arrays: only collect siblings of accessed fields
When an entire structure is accessed, users need to be able
to map those accesses to the members of the structure.
pet_scop_collect_arrays therefore recursively collects all subfields
of an outermost array.
However, being able to map a structure access to accesses
to its members is only important if one of those members
is also accessed individually.  Otherwise, the access
can simply be treated as an opaque access.
Modify pet_scop_collect_arrays to only collect siblings
of accessed members, rather than the entire member hierarchy.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>