pet_scop_collect_arrays: consider all accessed data spaces of an array
In particular, if there are any access relations associated to an array,
then also consider the data space in those access relations and not
just the data space of the index expression.
In the current code base, this change does not have any effect
since all subfields of an accessed arrays are recursively added
to the list of arrays, but a subsequent commit will change that
to only adding relevant subfields.
In order to be able to determine which subfields are relevant,
the code needs to consider the accesses in the access relations.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>