PetScan: extract out pet_expr_access_from_index
The PetScan::extract_access_expr method from which this function
is derived does not really use anything from the PetScan object.
It only uses the ASTContext that is stored in the object.
Pull it out of the PetScan class such that it can also be used
by the inliner in an upcoming commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>