configure.ac: drop -Wl,--no-keep-files-mapped from llvm-config --cxxflags output
Apparently, Ubuntu compiles clang with "-Wl,--no-keep-files-mapped",
which is a linker flag that is specific to gold, and it ends up
in the output of "llvm-config --cxxflags".  Now, "-fuse-ld=gold"
also appears in the output, but that flag apparently gets
stripped off by libtool, resulting in linker errors when
trying to link libpet.la if gold is not the default linker.
Arguably, these flags should not be included in the output
of "llvm-config --cxxflags", but remove "-Wl,--no-keep-files-mapped"
if it does.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>