only build isl.py and when extract_interface is available
The isl.py target in itself is fairly harmless since isl.py
is not built by default and therefore only results in failures
when the bindings are being built explicitly (or on "make dist").
However, the next commit will introduce a libdep.a target,
which well get built by default while
it can only be built with a bundled or build isl submodule.
Hide the entire isl.py machinery behind a conditional to avoid
failures when a system isl submodule is being used.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>