skip functions without valid scops when autodetecting
When autodetecting, PetScan::scan will return a NULL pointer
when it cannot find a valid scop in a given function.
This NULL pointer gets passed to call_fn, resulting in an error
being raised and no further functions being considered,
without any feedback to the user (since autodetection turn
off diagnostics).  This is especially problematic when
an unanalyzable function appears in an included header.
Simply ignore functions without a valid scop and move on
to the next function (if any).

Reported-by: Daniel Maier <daniel.maier@tu-berlin.de>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>