pet_expr_is_equal: compare index expressions using isl_multi_pw_aff_is_equal
Originally, the test was performed using isl_multi_pw_aff_plain_is_equal.
This was changed in pet-0.04-321-gbd0ad36 (pet_expr_is_equal: compare
index expressions up to reordering of parameters,
Thu Feb 20 10:40:36 2014 +0100) to allow for parameters appearing
in a different order.
It is not clear why isl_multi_pw_aff_is_equal was not used instead,
which allows for further differences (at a higher cost because
the affine expressions may internally need to be converted to a map).
Since pet_expr_is_equal is meant to check for equivalent expressions
(rather than identical expressions), it should allow more variation
in the index expression.
In fact, since the index expressions are relative to the expression domain,
in theory it should allow any pairs of expressions that are equivlant
on that domain, but this commit does not go that far.

The more relaxed comparison is useful for expressions such as

	[N, M] -> { [S_9[i] -> [i1]] -> [((0) : i1 = N - i;
		(1) : -i <= i1 < N - i; (0) : i1 = -1 - i)] }

This could also be represented as

	[N, M] -> { [S_9[i] -> [i1]] -> [((0) : i1 = N - i or
		i1 = -1 - i; (1) : -i <= i1 < N - i)] }

However, these are not considered to be obviously equivalent,
especially since the first gets more complicated domains
after parsing because the cells overlap.
These expressions are the result of a gist with respect to some domain.
Arguably, the overlapping cells make the first expression
an invalid representation of a piecewise expression,
even though the overlap only occurs outside of the domain
with respect to which the gist was taken.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>