preserve array extents in case of empty context
The context can be empty if it turns out the scop cannot
be executed for any value of the parameters.
Taking the gist with respect to such a context completely removes
any information about the array extents.
Even though these extents may not be strictly required
in such cases, there is no need to force the user
to add specific support for these cases.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>